pyproject.toml
[project]
name = "ai-vision-pipeline"
version = "0.1.0"
description = "Offline AI System for Human/Animal Detection and OCR"
authors = [
    {name = "Soumyajit Pal", email = "soumyajitpal416@gmail.com"},
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.29.0",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.2.0",
    "ultralytics>=8.1.0",
    "opencv-python-headless>=4.9.0",
    "paddlepaddle>=2.6.0",
    "paddleocr>=2.7.0",
    "loguru>=0.7.2",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0"
]
requires-python = ">=3.10"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "tests",
]
